[FLINK-14041][tests] Refactor LeaderRetrievalServiceHostnameResolutionTest
- rename class and test cases for clarity
- inline StandaloneUtils as it is just unnecessary
